+++ Suite.java
@@ -1,6 +1,6 @@
 public class Suite extends Room implements ComesWithDesignatedParking {
     private boolean hasBalcony;
-    private static final int BASE_PRICE_PER_NIGHT = 1200;
+    private static final double BASE_PRICE_PER_NIGHT = 1200.0;
 
     public Suite(String resID, int day, int month, int year,
                  Customer guest, int beds, String roomNumber,
@@ -12,15 +12,14 @@ public class Suite extends Room implements ComesWithDesignatedParking {
     public boolean isHasBalcony() { return hasBalcony; }
     public void setHasBalcony(boolean hasBalcony) { this.hasBalcony = hasBalcony; }
 
-    // Interface method — Suite gets a VIP parking spot
     @Override
     public void assignParking() {
         System.out.println("Suite " + roomNumber + " has been assigned a VIP parking spot.");
     }
 
     @Override
-    public int calculatePrice() {
-        int total = price * numberOfNights;
+    public double calculatePrice() {
+        double total = price * numberOfNights;
         if (hasBalcony) total += 300 * numberOfNights;
         return total;
     }
