+++ eventspace.java
@@ -2,7 +2,7 @@ public abstract class EventSpace extends Reservation {
     protected int maxCapacity;
     protected int sizeInSqFt;
 
-    public EventSpace(String resID, int day, int month, int year, int price,
+    public EventSpace(String resID, int day, int month, int year, double price,
                       Customer guest, int maxCapacity, int sizeInSqFt) {
         super(resID, day, month, year, price, guest);
         this.maxCapacity = maxCapacity;
@@ -16,5 +16,5 @@ public abstract class EventSpace extends Reservation {
     public void setSizeInSqFt(int sizeInSqFt) { this.sizeInSqFt = sizeInSqFt; }
 
     @Override
-    public abstract int calculatePrice();
+    public abstract double calculatePrice();
 }
